feat(shortcuts): removed binding of shortcuts out of the player
An object is now "actionable" which means it exports actions that can
be executed by some kind of controller. This controller can i.e. execute
an action based on keydown events (shortcuts).
Currently it's only static shortcuts but this system allow for rebinding
the keys easily as the player doesn't care how a shortcut is executed.